Remove hard-js9 dependency during basic sal import
